Clear static state between tests
BuildLoopMonitorRepository caches a BuildLoopMonitor in static state which may
be shared between all of the tests in DashboardControllerTest.  The cached
BuildLoopMonitor can be purged by calling cancelPosting on the repository.

Unfortunately, testShouldBeAbleToStartPostingIfThereIsNOExistingPosting is
dependent on not having any BuildLoopMonitor cached at the beginning of the
test.  Instead of merely forcing the BuildLoopMonitorRepository to purge its
cache at the beginning of that test, let's do that in setUp() so that every
test starts with a clean repository.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>