Copy XPP and XStream into binary release zip
Make sure we explicitly copy both jars into the binary release zip because
they live in the shared lib/ directory.  This temporarily means there will
be duplicate copies in both the source tree as well as cruisecontrol.jar and
dashboard.war.  I plan on fixing that soon, but in the meantime this is
necessary for upgrading XStream to work with JDK 7.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>