Hook latest XPP and XStream to the build
Actually upgrade our dependency to now point at the latest versions of both
XPP and XStream.  Remove old versions of both jar files.

This is necessary to build with JDK 7.  We don't actually use any new Java 7
features or APIs, but due to Oracle's acquisition of Sun Microsystems some
internal strings changed which caused old versions of XStream to incorrectly
use the wrong reflection engine (see: http://jira.codehaus.org/browse/XSTR-651)
which caused a number of test failures.

I verified that we still build successfully with JDK 5 and JDK 6 using the
latest versions of these libraries.

Also, be note that just because we are compatible with JDK 7 does not mean it
is ok to introduce any dependencies on Java 7.  It is expected that we remain
fully source and API compatible with Java 5.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>