Initial IntelliJ IDEA configuration
Generated with IntelliJ IDEA 13; might not be compatible with earlier versions.
We already have Eclipse .classpath and .project in the tree, and since IntelliJ
now offers a free Community Edition I think it makes sense to also add IntelliJ
configuration in the tree.  Hopefully, this will make it easier for others to
setup a CruiseControl development environment and contribute to the project, as
well as make it easier for existing developers to migrate to new machines.  :)

Currently, there are only modules for .ant-scripts/, main/ and main/loggers/.
I'm still working through the library dependencies necessary for the reporting
app and dashboard.  Also, I haven't tried running the functional tests in main/
from within IntelliJ yet, so there might be some more tweaks necessary there.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>