Configuration handling rewrite.
dist: the directory where the CC was installed (contains directories like etc/, lib/, webapps/, etc.)
- if <dist>...</dist>  (or -dist option) directory is set but invalid, CC fails
  if it is not set, it is guessed as the parent dir of lib/cruisecontrol-launcher.jar, which should work on normal configuration ...

proj: the directory where the CC projects are located (contains directories like projects/, log/, artifacts/, etc.)
- if <proj>...</proj>  (or -proj option) directory is set but invalid, CC fails
  if it is not set, it is guessed as the current working directory

lib: the list of directories where to find the libraries required for cruisecontrol run
- for non absolute paths in <lib>...</lib>  (or -lib option), the ${dist}/${lib} path is used
  if not set, ${dist}/lib path is used
- there may be several lib directories set, either by repeating <lib>...</lib> XML item, or by repeating -lib option. When set as property,
  the paths must be separated by the valid path separator.
  The cruisecontrol-launcher always includes libraries from the directory where it was started; this directory does not have to be set though

user_lib: the list of directories where to find additional libraries, e.g. with 3rd party modules
- for non absolute paths in <user_lib>...</user_lib>  (or -user_lib option), the ${proj}/${lib} path is used
- in addition, the /home/user/.cruisecontrol/lib directory is searched as well (does not have to be configured)

home:
- this option was removed and replaced by proj (better matches to what it represents)

The dist, proj and lib options are (to some extend) tried to determine automatically, so it should be possible to run the cruisecontrol even with
almost empty <launch> configuration, if installed as packaged. The example of the working configuration is in binaryrelease/launch.xml