Configuration refactoring:
The Main and Launcher were originally decoupled, joined through CruiseControlMain interface only. When the Configuration was inroduced, this
decoupling was broken (there is complex Configuration object passed from Launcher to Main), and thus both cruisecontrol.java and cruisecontrol-launcher.java
mst be compiled from the same version of the repository, which was no the original design idea I believe.

In this commit, the decupling was restored again, yet the CruiseControlMain interface was extended to pass the configurarion to the main cruiseconrol
process.

Launcher is now resposible for the most basic settings such as merging the configuration from a config file, properties and command line arguments, and
then for determining library, working and project directories, the main cruisecontrol configuration file and setting log4j config. All the other options
are passed to the main CruiseControl object.

The configuration options passed to the main cruisecontrol process are than used to configure the main cruisecontrl process.

All the changes required the following modifications:
- extension of the CruiseControlMain
- new Configuration interface allowing to pass the config from Launcher to (CruiseControl) Main
- split of config options (some are known o Launcher, the remaining to the Main
- test tweaks and extensons

Otherwise, the configuration options remain the same as before. The change affects implementation but not the usage.

There is a help printing left unfinished ...